Fixed calculations.

+++ trunk/src/com/thoughtworks/billing/service/Bill.java
@@ -2,14 +2,13 @@ package com.thoughtworks.billing.service;
 
 import com.thoughtworks.billing.bean.Item;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public interface Bill {
     List<Item> listSoldItems();
 
-    BigDecimal getTotalTax();
+    double getTotalTax();
 
-    BigDecimal getTotalPrice();
+    double getTotalPrice();
 
 }
